requirements.txt
azure-functions
pymongo
requests
python-dotenv
instagram-api-python